refactor: split basic.test file
